federal/congressional/federal_ingest_congressional_hearings/requirements.txt
google-cloud-secret-manager==2.2.0
google-cloud-firestore==2.0.2
requests>=2.25.1
elasticsearch>=7.0.0,<8.0.0
beautifulsoup4==4.9.3
pandas==1.2.5
xmltodict==0.12.0
furl==2.1.2
